+++ CHANGELOG.md
@@ -2,13 +2,14 @@
 
 **1.7.0**
 * Update JPA api to 2.2
-* Support plugin JAX-RS #448
+* DB: add annotation to mark created and last modified timestamp fields #480
 * Missing actframework version in the banner text in dev mode with Eclipse project #478
 * Rename `AppEvent` to `SysEvent` #475
 * Improve simple event handler mechanism #473
 * Emit/trigger `EventObject` typed event through `EventBus #472
 * SimpleEventHandler mechanism does not work with `EventObject` #471
 * Allow `LoadResource` inject into Set of strings #470
+* Support plugin JAX-RS #448
 
 **1.6.4 14/Jan/2018**
 * Performance issue with Rythm #469